+++ Examples/Events Demo/Classes/iRateAppDelegate.m
@@ -59,12 +59,6 @@
 #pragma mark -
 #pragma mark Memory management
 
-- (void)dealloc
-{
-    [viewController release];
-    [window release];
-    [super dealloc];
-}
 
 
 @end