+++ Examples/Mac Demo/iRateMacAppDelegate.h
@@ -10,9 +10,9 @@
 
 @interface iRateMacAppDelegate : NSObject <NSApplicationDelegate>
 {
-    NSWindow *window;
+    NSWindow *__unsafe_unretained window;
 }
 
-@property (assign) IBOutlet NSWindow *window;
+@property (unsafe_unretained) IBOutlet NSWindow *window;
 
 @end