+++ Examples/Zero Config Demo/Classes/iRateAppDelegate.h
@@ -12,8 +12,8 @@
 
 @interface iRateAppDelegate : NSObject <UIApplicationDelegate>
 
-@property (nonatomic, retain) IBOutlet UIWindow *window;
-@property (nonatomic, retain) IBOutlet iRateViewController *viewController;
+@property (nonatomic, strong) IBOutlet UIWindow *window;
+@property (nonatomic, strong) IBOutlet iRateViewController *viewController;
 
 @end