+++ Examples/iPhone Demo/Classes/iRateAppDelegate.m
@@ -46,12 +46,6 @@
 #pragma mark -
 #pragma mark Memory management
 
-- (void)dealloc
-{
-    [viewController release];
-    [window release];
-    [super dealloc];
-}
 
 
 @end