+++ iRate/iRate.h
@@ -1,7 +1,7 @@
 //
 //  iRate.h
 //
-//  Version 1.7 beta
+//  Version 1.7
 //
 //  Created by Nick Lockwood on 26/01/2011.
 //  Copyright 2011 Charcoal Design
@@ -83,6 +83,7 @@ iRateErrorCode;
 - (void)iRateUserDidAttemptToRateApp;
 - (void)iRateUserDidDeclineToRateApp;
 - (void)iRateUserDidRequestReminderToRateApp;
+- (BOOL)iRateShouldOpenAppStore;
 
 @end