+++ RELEASE NOTES.md
@@ -1,7 +1,8 @@
-Version 1.7 beta
+Version 1.7
 
-- On iOS 6, iRate can now use the StoreKit APIs to display the product page directly within the app. To enable this feature you will need to add the StoreKit framework to your project.
-- Moved the ARCHelper macros out of the header file so that they do not affect other code in the project that is not ARC compliant
+- On iOS 6, iRate can now use the StoreKit APIs to display the product page directly within the app.
+- iRate now requires the Storekit framework on iOS
+- iRate now requires ARC. To use iRate in a non-ARC project, follow the instructions in the README file.
 - Dropped support for 32-bit Macs running Snow Leopard
 - Added Swedish translation